Перенес блокировку форм в утилиты
