+++ source/js/util.js
@@ -16,7 +16,29 @@ function debounce(callback, delay = 500) {
   }
 }
 
+// Снять блокировку элементов формы
+function enableForm(formData, classData) {
+  formData.classList.remove(classData);
+  const formElements = formData.querySelectorAll('input, select, textarea, button');
+
+  formElements.forEach(function (formElement) {
+    formElement.removeAttribute('disabled');
+  });
+}
+
+// Добавить блокировку элементов формы
+function disableForm(formData, classData) {
+  formData.classList.add(classData);
+  const formElements = formData.querySelectorAll('input, select, textarea, button');
+
+  formElements.forEach(function (formElement) {
+    formElement.setAttribute('disabled', 'disabled');
+  });
+}
+
 export {
   isEscEvent,
-  debounce
+  debounce,
+  enableForm,
+  disableForm
 };
